home udemy practice 03/01/16

+++ Section10/src/InnerClasses/GearBox.java
@@ -19,7 +19,7 @@ public class GearBox {
       this.clutchIsIn = in;
    }
 
-   public void addGear(int number, int ratio) {
+   public void addGear(int number, double ratio) {
       if ((number > 0) && number <= this.maxGears) {
          this.gears.add(new Gear(number, ratio));
       }
